New features: ============= The Gaudi and GaudiPython applications have a new attribute called 'setupProjectOptions'. It contains extra options to be passed onto the SetupProject command used for configuring the environment. As an example setting it to '--dev' will give access to the DEV area. For full documentation of the available options see https://twiki.cern.ch/twiki/bin/view/LHCb/SetupProject. The 'lhcb_release_area' attribute has been taken away as it was not useful.
The Gaudi and GaudiPython applications can now read data from the
detector. For this a new attribute, 'datatype_string', is added to the
LHCbDataset. It contains the string that is added after the filename
in the options to tell Gaudi how to read the data. If reading raw data
(mdf files) it should be set to "SVC='LHCb::MDFSelector'".

Minor changes:
==============
The identification of which default application version to pick is now
using SetupProject.

Many test cases have been updated to Ganga 5.