Remove usage of the 'print' statement
In most cases I replaced it with logger.info() (or occasionally warning or error).

There are some places where it doesn't make sense to use the logger (writing
to files) so I used the print function there. Many of these uses could be replaced
with file.write() type calls.

There are also a number of print statements left in the code. A lot of these are
in modules which I think will eb removed soon anyway.