Use `in dir()` rather than `hasattr()` in `stripProxy()`
`hasattr()` calls `getattr()` and checks to see if an exception is
raised. In some situations, `getattr()` calls `stripProxy()` which
causes a recursive loop. This was causing errors in the test suite like:

```
Exception RuntimeError: 'maximum recursion depth exceeded while calling a Python object' in <type 'exceptions.AttributeError'> ignored
```

I'm not sure that `in dir()` is definitely the best option but it serves
the purpose. `in obj.__dict__` does not work as built-in types like
`str` do not have the `__dict__` attribute.