Add an optional `create` flag to `getConfig`
By default at present getConfig will create a config section if one was not
found. This change maintains that behaviour but allows for it to be turned
off by setting `create` to `False`.

In future I'd like to change this behaviour so that it always raises an
exception if the config section was not found.