Add thread-safe locks around schema access and parent access in GangaObject
Every GangaObject has a _lock attribute which is a threading.RLock. This is
accessed through the lock property on the object which as well as locking
the object, also locks all its descendants. The Descriptor's __get__ and
__set__ use this lock to ensure only one thread can access the object's
schema attributes at any time.

Some code will need to claim explicit locks which can be done by using the
lock property as a context manager.