Make updateDict_ts so that it's recreated each time
When stopping and restarting the monitoring, some jobs were not being
monitored as the system had stale data claiming that they were already
covered and so no new monitoring should be performed.

By making updateDict_ts a member of the JobRegistry_Monitor class it will
be recreated every time the monitoring is started and so will recheck
everything.