Make GangaObjects only lock their root object
Rather than having a lock on every single GangaObject which then propagates
up the parent chain, instead just acquire the lock in the object's root
object. This is sensitive to altering the parent of an object so extra
logic is need in _setParent to lock both the old and new root.