Cache the LHCb DIRAC environment in a JSON file
This avoids round tripping of source a command → save it to a file → source
it again → use it. The environment is still cached per session in memory.

Other plugins can do the same but only LHCb are slow enough to need it. The
code to do the caching is moved to the DIRAC plugin.

Deprecate the `DiracEnvFile` option and replace it with a `DiracEnvSource`
option which should point to a sourceable script to provide the
environment.

`DiracEnvJSON` can point to a specific cached env to override the source
command.