Switch to shell style ENTRYPOINT
using the recommended list style was not allowing us to execute multiple commands (at least I couldnt get it to ;-) ). Therefore switch to shell style and accept the consequences. Also run the CVMFS and login scripts manually as overriding the ENTRYPOINT from base image. Note using CMD as previous commit did not work as these are considered arguments to the base layers ENTRYPOINT.

	modified:   python/GangaLHCb/test/Dockerfile