Update Notifier.py
Add in the notification mail from ganga also the job name.
It would be nice to add as well in the mail sent the user which has done that in such a way if the notifier sends mails to a MailingList of a shared analysis project one can know which one has completed the job. 
I am just not sure how to get the ```whoami``` inside ganga libraries.